fix: fix some issues
