+++ src/utils/defaultInputsData.js
@@ -1,0 +1,17 @@
+const defaultRegisterData = {
+  directionId: '',
+  name: '',
+  email: '',
+  lastName: '',
+  sex: '',
+  education: '',
+  birthDate: '',
+  universityAverageScore: '',
+  mathScore: '',
+  address: '',
+  mobilePhone: '',
+  skype: '',
+  startDate: '',
+};
+
+export { defaultRegisterData };
